+++ src/main/java/com/group/libraryapp/domain/user/UserRepository.java
@@ -1,5 +1,6 @@
 package com.group.libraryapp.domain.user;
 
+import com.group.libraryapp.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
