+++ src/main/java/com/group/libraryapp/dto/user/response/UserResponse.java
@@ -1,6 +1,6 @@
 package com.group.libraryapp.dto.user.response;
 
-import com.group.libraryapp.domain.user.User;
+import com.group.libraryapp.domain.User;
 
 public class UserResponse {
