+++ src/main/kotlin/com/group/libraryapp/domain/User.kt
@@ -1,0 +1,41 @@
+package com.group.libraryapp.domain
+
+import javax.persistence.CascadeType
+import javax.persistence.Entity
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+import javax.persistence.OneToMany
+
+@Entity
+class User(
+    var name: String,
+
+    var age: Int?,
+
+    @OneToMany(mappedBy = "user", cascade = [CascadeType.ALL], orphanRemoval = true)
+    val userLoanHistories: MutableList<UserLoanHistory> = mutableListOf(),
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long? = null
+) {
+
+    init {
+        if (name.isBlank()) {
+            throw IllegalStateException("이름은 비어 있을 수 없습니다")
+        }
+    }
+
+    fun updateName(name: String) {
+        this.name = name
+    }
+
+    fun loanBook(book: Book) {
+        this.userLoanHistories.add(UserLoanHistory(this, book.name, false))
+    }
+
+    fun returnBook(bookName: String) {
+        this.userLoanHistories.first { history -> history.bookName == bookName }.doReturn()
+    }
+}
